added the addEventListeners method for document.body so we can use it for the handleFormClick method. Also added it to the constructor so it runs when the app is created.
Using query selector we reference the form, noteTitle and formbuttons html id elements in the constuctor.

on the handleFormClick we store the true/false value of the event.target to check if there is a click. when true the new note form expanse bc we add the form-open class and be display as block the noteTitle and the formButtons. when false remove the open class and set the dsplay of buttons and title to none.